Require atomic swap contracts to specify the secret size.
This size is checked during audit time to ensure that the secret is of
an expected size and will not be larger than usable by one of the
networks.  This prevents a fraud attack where an attacker is able to
redeem a contract using a secret that is larger than the other network
is capable of using.